Extract configuration to break dependency cycle
