+++ src/function-builder.ts
@@ -20,9 +20,17 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-import * as _ from 'lodash';
 import * as express from 'express';
+import * as _ from 'lodash';
 
+import { CloudFunction, EventContext } from './cloud-functions';
+import {
+  DeploymentOptions,
+  MAX_TIMEOUT_SECONDS,
+  RuntimeOptions,
+  SUPPORTED_REGIONS,
+  VALID_MEMORY_OPTIONS,
+} from './function-configuration';
 import * as analytics from './providers/analytics';
 import * as auth from './providers/auth';
 import * as crashlytics from './providers/crashlytics';
@@ -32,35 +40,6 @@ import * as https from './providers/https';
 import * as pubsub from './providers/pubsub';
 import * as remoteConfig from './providers/remoteConfig';
 import * as storage from './providers/storage';
-import { CloudFunction, EventContext, Schedule } from './cloud-functions';
-
-/**
- * List of all regions supported by Cloud Functions.
- */
-const SUPPORTED_REGIONS = [
-  'us-central1',
-  'us-east1',
-  'us-east4',
-  'europe-west1',
-  'europe-west2',
-  'asia-east2',
-  'asia-northeast1',
-];
-
-/**
- * List of available memory options supported by Cloud Functions.
- */
-const VALID_MEMORY_OPTS = ['128MB', '256MB', '512MB', '1GB', '2GB'];
-
-// Adding this memory type here to error on compile for TS users.
-// Unfortunately I have not found a way to merge this with VALID_MEMORY_OPS
-// without it being super ugly. But here they are right next to each other at least.
-type Memory = '128MB' | '256MB' | '512MB' | '1GB' | '2GB';
-
-/**
- * Cloud Functions max timeout value.
- */
-const MAX_TIMEOUT_SECONDS = 540;
 
 /**
  * Assert that the runtime options passed in are valid.
@@ -70,10 +49,10 @@ const MAX_TIMEOUT_SECONDS = 540;
 function assertRuntimeOptionsValid(runtimeOptions: RuntimeOptions): boolean {
   if (
     runtimeOptions.memory &&
-    !_.includes(VALID_MEMORY_OPTS, runtimeOptions.memory)
+    !_.includes(VALID_MEMORY_OPTIONS, runtimeOptions.memory)
   ) {
     throw new Error(
-      `The only valid memory allocation values are: ${VALID_MEMORY_OPTS.join(
+      `The only valid memory allocation values are: ${VALID_MEMORY_OPTIONS.join(
         ', '
       )}`
     );
@@ -128,18 +107,6 @@ export function runWith(runtimeOptions: RuntimeOptions): FunctionBuilder {
   if (assertRuntimeOptionsValid(runtimeOptions)) {
     return new FunctionBuilder(runtimeOptions);
   }
-}
-
-export interface RuntimeOptions {
-  timeoutSeconds?: number;
-  memory?: Memory;
-}
-
-export interface DeploymentOptions {
-  regions?: string[];
-  timeoutSeconds?: number;
-  memory?: Memory;
-  schedule?: Schedule;
 }
 
 export class FunctionBuilder {
