+++ src/index.ts
@@ -23,8 +23,6 @@
 // Providers:
 import * as analytics from './providers/analytics';
 import * as auth from './providers/auth';
-
-import * as apps from './apps';
 import * as crashlytics from './providers/crashlytics';
 import * as database from './providers/database';
 import * as firestore from './providers/firestore';
@@ -32,6 +30,8 @@ import * as https from './providers/https';
 import * as pubsub from './providers/pubsub';
 import * as remoteConfig from './providers/remoteConfig';
 import * as storage from './providers/storage';
+
+import * as apps from './apps';
 import { handler } from './handler-builder';
 import { setup } from './setup';
 
@@ -52,8 +52,9 @@ export {
 };
 
 // Exported root types:
-export * from './config';
 export * from './cloud-functions';
+export * from './config';
 export * from './function-builder';
+export * from './function-configuration';
 
 setup();
