+++ src/providers/crashlytics.ts
@@ -21,11 +21,11 @@
 // SOFTWARE.
 
 import {
-  makeCloudFunction,
   CloudFunction,
   EventContext,
+  makeCloudFunction,
 } from '../cloud-functions';
-import { DeploymentOptions } from '../function-builder';
+import { DeploymentOptions } from '../function-configuration';
 
 /** @internal */
 export const provider = 'google.firebase.crashlytics';
