+++ src/providers/remoteConfig.ts
@@ -28,7 +28,7 @@ import {
   EventContext,
   makeCloudFunction,
 } from '../cloud-functions';
-import { DeploymentOptions } from '../function-builder';
+import { DeploymentOptions } from '../function-configuration';
 
 /** @internal */
 export const provider = 'google.firebase.remoteconfig';
